srodolfobarbosa/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
loguru==0.7.0
python-dotenv==1.0.0
langchain-groq==0.1.0
supabase==1.0.0
pinecone-client>=2.2.1,<4.0.0
mercadopago>=2.3.0,<3.0.0
duckduckgo-search==2.9.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
httpx==0.23.3
pypdf2>=3.0.1,<4.0.0
Pillow==10.0.0
pgvector==0.4.0
huggingface_hub
requests